Move generating datasets to separate file
