+++ src/App.vue
@@ -53,49 +53,7 @@
 
 import Chart from './components/Chart.vue'
 import {computeTruePositives, computeFalsePositives, computeTrueNegatives, computeFalseNegatives} from "./metrics";
-import {datasets} from '../assets/datasets/all.json'
-
-const k = 10;
-const n = 50;
-
-function range(size, startAt = 0) {
-    return [...Array(size).keys()].map(i => i + startAt);
-}
-
-const tileSize = 1;
-
-const tiles = range(k * k)
-  .map((i) => {
-    return {
-      x: i % k,
-      y: Math.floor(i / k),
-      v: 0 // Math.random()
-    };
-  });
-
-const points = range(n)
-  .map(() => {
-    const x = Math.random();
-    const y = Math.random();
-
-    return {
-      x: k * x,
-      y: k * y,
-      v: +(y < Math.sin(4 * x))
-    };
-  });
-
-const circle = range(n)
-  .map(() => {
-    const x = Math.random();
-    const y = Math.random();
-
-    return {
-      x: k * x,
-      y: k * y,
-      v: +(Math.pow(x - 0.5, 2) + Math.pow(y - 0.5, 2) < 0.1)
-    };
-  });
+import {tiles, tileSize, generatedDatasets, fixedDatasets} from './datasets.js'
 
 export default {
   name: 'App',
@@ -107,11 +65,7 @@ export default {
       dataset: {name: "Empty", points: []},
       points: [],
       tiles: tiles,
-      options: datasets.map(
-        (dataset) => ({
-          name: dataset.name,
-          points: dataset.data.train
-        })),
+      options: generatedDatasets,
     };
   },
   computed: {
